requirements.txt
aiohttp==3.8.5
aiosignal==1.3.2
anyio==4.7.0
astral==2.2
async-timeout==4.0.2
atomicwrites-homeassistant==1.4.1
attrs==22.2.0
awesomeversion==22.9.0
bcrypt==4.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
ciso8601==2.3.0
cryptography==41.0.1
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
home-assistant-bluetooth==1.10.0
homeassistant==2023.7.3
httpcore==0.17.3
httpx==0.24.1
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.2
lru-dict==1.2.0
MarkupSafe==3.0.2
multidict==6.1.0
orjson==3.9.1
pycparser==2.22
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-slugify==4.0.1
pytz==2024.2
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
text-unidecode==1.3
typing_extensions==4.12.2
ulid-transform==0.7.2
urllib3==2.2.3
voluptuous==0.13.1
voluptuous-serialize==2.6.0
yarl==1.9.2
